% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_features_weather_30min.R
\name{calc_features_weather}
\alias{calc_features_weather}
\title{Calculates features from one environmental time-series variable and smart meter data}
\usage{
calc_features_weather(SMD, WEATHER, rowname = NULL)
}
\arguments{
\item{SMD}{the load trace for one week (vector with 672 or 336 elements)}

\item{WEATHER}{weather observations (e.g. temperature) in 30-minute readings
(vector with 336 elements)}

\item{rowname}{the row name of the current data point}
}
\description{
Calculates features from one environmental time-series variable and smart meter data
}
\references{
Hopf, K. (2019). Predictive Analytics for Energy Efficiency and
    Energy Retailing (1st ed.). Bamberg: University of Bamberg.
    \doi{10.20378/irbo-54833}

Hopf, K., Sodenkamp, M., Kozlovskiy, I., & Staake, T. (2014).
    Feature extraction and filtering for household classification based on
    smart electricity meter data. Computer Science-Research and Development,
    (31) 3, 141–148. \doi{10.1007/s00450-014-0294-4}

Hopf, K., Sodenkamp, M., & Staake, T. (2018). Enhancing energy
    efficiency in the residential sector with smart meter data analytics.
    Electronic Markets, 28(4). \doi{10.1007/s12525-018-0290-9}
}
\author{
Konstantin Hopf \email{konstantin.hopf@uni-bamberg.de}, Ilya Kozlovslkiy
}
