% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_features_daily_htnt2.R
\name{calc_featureshtnt_consumption2}
\alias{calc_featureshtnt_consumption2}
\title{Calculates consumption features from daily (HT / NT) smart meter data}
\usage{
calc_featureshtnt_consumption2(
  HTCons,
  NTCons,
  rowname = NULL,
  featsCoarserGranularity = FALSE
)
}
\arguments{
\item{HTCons}{a vector with 7 measurements for HT consumption in one week (beginning with monday)}

\item{NTCons}{a vector with 7 measurements for NT consumption in one week (beginning with monday)}

\item{rowname}{the row name of the resulting feature vector}

\item{featsCoarserGranularity}{are the features of finer granularity levels also to be calculated (T/FALSE)}
}
\description{
The division in HT / NT is done from the input smart meter data
}
\references{
Hopf, K. (2019). Predictive Analytics for Energy Efficiency and
    Energy Retailing (1st ed.). Bamberg: University of Bamberg.
    \doi{10.20378/irbo-54833}
}
\author{
Konstantin Hopf \email{konstantin.hopf@uni-bamberg.de}
}
