% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_encode_p_val_stars.R
\name{encode_p_val_stars}
\alias{encode_p_val_stars}
\title{Encodes p-values with a star rating according to the Significance code:}
\usage{
encode_p_val_stars(pval)
}
\arguments{
\item{pval}{the p-value}
}
\value{
character with the encoding
}
\description{
'.' for p-value < 0.1, '*' for < 0.05, '**' for < 0.01, '***' for < 0.001
}
\author{
Konstantin Hopf \email{konstantin.hopf@uni-bamberg.de}
}
