% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_all_subsets.R
\name{features_all_subsets}
\alias{features_all_subsets}
\title{Creates a set of all combinations of features}
\usage{
features_all_subsets(set)
}
\arguments{
\item{set}{vector of available festures that are premutated}
}
\value{
a list of subsets of the input vector
}
\description{
Creates a set of all combinations of features
}
\examples{
    features_all_subsets(c("A", "B", "C"))
}
\author{
Konstantin Hopf \email{konstantin.hopf@uni-bamberg.de}, Ilya Kozlovskiy
}
