% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_getDay_US_week.R
\name{getDay_US_week}
\alias{getDay_US_week}
\title{Retrieves the date of the monday in a US week-string (as implemented by R as.Date)}
\usage{
getDay_US_week(
  theweek,
  day = c("Mon", "Tue", "Wed", "Thu", "Fri", "Sat", "Sun")
)
}
\arguments{
\item{theweek}{the string with the week name}

\item{day}{the weekday that shall be returned}
}
\value{
the date of the weekday in the given week
}
\description{
According to date formats defined by ISO 8601:
 * Single days are written in yyy-mm-dd (y: year, m: month, d: day); e.g., 2016-07-19
 * Weeks are written in yyyy-WUww; e.g., 2016-WU29 (typically with the first Sunday of the year as day 1 of week 1)
}
\author{
Konstantin Hopf \email{konstantin.hopf@uni-bamberg.de}
}
