% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_NA_Inf_omit.R
\name{naInf_omit}
\alias{naInf_omit}
\title{Removes the rows with \link{NA} or \link{Inf} values}
\usage{
naInf_omit(V)
}
\arguments{
\item{V}{A \link{data.frame} or \link{matrix} which has to be cleaned}
}
\value{
A cleaned version of \link{data.frame} or \link{matrix}
}
\description{
Cleans up a \link{data.frame} or \link{matrix} which is useful for
cases wehere you need complete datasets
}
\seealso{
\link{replaceNAsFeatures}, \link{remove_empty_features}
}
\author{
Konstantin Hopf \email{konstantin.hopf@uni-bamberg.de}
}
