% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_occupancy_cluster.R
\name{occupancy_cluster}
\alias{occupancy_cluster}
\title{Determines two clusters of high and low consumption times (e.g., non-ocupancy during holidays)}
\usage{
occupancy_cluster(consumption, n_days_check = 4, sds_between_clusters = 1.5)
}
\arguments{
\item{consumption}{the consumption time series}

\item{n_days_check}{number of consecutive days that should be considered as a minimal cluster}

\item{sds_between_clusters}{the multiples of standatd deviation that must be at least between the cluster centers (decimal number)}
}
\value{
list with cluster assignments and the k-Means clustering model
}
\description{
Determines two clusters of high and low consumption times (e.g., non-ocupancy during holidays)
}
\author{
Konstantin Hopf \email{konstantin.hopf@uni-bamberg.de}
}
