% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_replaceNAsFeatures.R
\name{replaceNAsFeatures}
\alias{replaceNAsFeatures}
\title{Replaces NA values with a given ones}
\usage{
replaceNAsFeatures(indata, features, replacement = 0)
}
\arguments{
\item{indata}{a \link{data.frame}}

\item{features}{a vector of variable names (must be colum names of \code{indata} that are
to be used for \link{NA}-replacement}

\item{replacement}{the alternative value, \link{NA} values should be replaced with, zero by default}
}
\value{
the modified data.frame with replaced values
}
\description{
Taks a \link{data.frame} and replaces all \link{NA} values with a certain value.
}
\seealso{
\link{naInf_omit}, \link{remove_empty_features}
}
\author{
Konstantin Hopf \email{konstantin.hopf@uni-bamberg.de}
}
